Feature validation

+++ src/main/resources/static/js/main.js
@@ -33,6 +33,12 @@ let getAll = () => {
 
 // Get By ID
 let getByID = () => {
+
+    if (!idValidation()) {
+        alert("An ID needs to be entered.")
+        return;
+    }
+
     axios.get(`http://localhost:8080/student/getById/${inputID.value}`)
         .then(res => {
             console.log(res.data);
@@ -41,6 +47,12 @@ let getByID = () => {
 
 // Get By First Name
 let getByFirstName = () => {
+
+    if (!firstNameValidation()) {
+        alert("A first name needs to be entered.")
+        return;
+    }
+
     axios.get(`http://localhost:8080/student/getByFirstName/${inputFirstName.value}`)
         .then(res => {
             console.log(res.data);
@@ -49,6 +61,12 @@ let getByFirstName = () => {
 
 // Get Second Name
 let getBySecondName = () => {
+
+    if (!secondNameValidation()) {
+        alert("A second name needs to be entered.")
+        return;
+    }
+
     axios.get(`http://localhost:8080/student/getBySecondName/${inputSecondName.value}`)
         .then(res => {
             console.log(res.data);
@@ -57,6 +75,12 @@ let getBySecondName = () => {
 
 // Get By Age
 let getByAge = () => {
+
+    if (!ageValidation()) {
+        alert("An age needs to be entered.")
+        return;
+    }
+
     axios.get(`http://localhost:8080/student/getByAge/${inputAge.value}`)
         .then(res => {
             console.log(res.data);
@@ -65,6 +89,11 @@ let getByAge = () => {
 
 // Create
 let create = () => {
+
+    if (!createValidation()) {
+        alert("All required fields do not have values.")
+        return;
+    }
 
     let obj = {
         "firstName": inputFirstName.value,
@@ -83,6 +112,12 @@ let create = () => {
 
 // Update
 let update = () => {
+
+    if (!updateValidation()) {
+        alert("All required fields do not have values.")
+        return;
+    }
+
     let obj = {
         "firstName": inputFirstName.value,
         "secondName": inputSecondName.value,
@@ -99,7 +134,14 @@ let update = () => {
 }
 
 // Delete
+
 let del = () => {
+
+    if (!deleteValidation()) {
+        alert("An ID is needed to carry out this operation.")
+        return;
+    }
+
     axios.delete(`http://localhost:8080/student/delete/${inputID.value}`)
         .then(res => {
             console.log(res.data);
@@ -116,3 +158,75 @@ getByIDBtn.addEventListener("click", getByID);
 getByFirstNameBtn.addEventListener("click", getByFirstName);
 getBySecondNameBtn.addEventListener("click", getBySecondName);
 getByAgeBtn.addEventListener("click", getByAge);
+
+// VALIDATION
+// Get By ID
+let idValidation = () => {
+    if (inputID.value === "") {
+        return false;
+    } else {
+        return true;
+    }
+}
+
+// Get By First Name
+let firstNameValidation = () => {
+    if (inputFirstName.value === "") {
+        return false;
+    } else {
+        return true;
+    }
+}
+
+// Get By Second Name   
+let secondNameValidation = () => {
+    if (inputSecondName.value === "") {
+        return false;
+    } else {
+        return true;
+    }
+}
+
+// Get By Age
+let ageValidation = () => {
+    if (inputAge.value === "") {
+        return false;
+    } else {
+        return true;
+    }
+}
+
+// Create
+let createValidation = () => {
+    if (inputFirstName.value === "" || inputSecondName.value === ""
+        || inputAge.value === "" || inputContactNumber.value === ""
+        || inputEmail.value === "") {
+        return false;
+    } else {
+        return true;
+    }
+}
+
+// Update
+let updateValidation = () => {
+    if (inputID.value === "" || inputFirstName.value === "" || inputSecondName.value === ""
+        || inputAge.value === "" || inputContactNumber.value === ""
+        || inputEmail.value === "") {
+        return false;
+    } else {
+        return true;
+    }
+}
+
+// Delete
+let deleteValidation = () => {
+    if (inputID.value === "") {
+        return false;
+    } else {
+        return true;
+    }
+}
+
+
+
+
